# spec/api.yaml
...
openapi: 3.0.3
info:
  title: TUD Assistance Backbone API
  description: Component for analyzing learning process data in the form of xAPI statements and providing feedback to
    learners' interactions and assistance and suggestions with regard to the user's learning state and the corresponding
    learning content.
  license:
    name: GNU General Public License
    url: https://www.gnu.org/licenses/old-licenses/gpl-3.0.de.html
  version: 1.0.0
...
security:
- bearerAuth: []
paths:
...
  /api/v1/student:
    get:
      tags:
      - Student Module
      summary: Get Student Models
      description: Get all Student Models.
      operationId: getStudentModels
      parameters:
      - $ref: '#/components/parameters/pageNumber'
      - $ref: '#/components/parameters/numberOfObjectsPerPage'
      responses:
        200:
          description: Successfully retrieved Student Models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentModelList'
        400:
          description: Bad request
        401:
          description: Authentication failed
  /api/v1/student/{userId}:
    get:
      tags:
      - Student Module
      summary: Get student model
      description: Get the student model for a certain user with the specified ID.
      operationId: getStudentModel
      parameters:
      - $ref: '#/components/parameters/userId'
      responses:
        200:
          description: Successfully retrieved student model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentModel'
        400:
          description: Bad request
        401:
          description: Authentication failed
        404:
          description: User not found
    delete:
      tags:
      - Student Module
      summary: Reset student model
      description: Reset the student model for a certain user with the specified ID.
      operationId: resetStudentModel
      parameters:
      - $ref: '#/components/parameters/userId'
      responses:
        204:
          description: Successfully reset student model
        400:
          description: Bad request
        401:
          description: Authentication failed
        404:
          description: User not found
  /api/v1/student/progress:
    post:
      tags:
      - Student Module
      summary: Search for student progress
      description: Search the progress for a specific user with the specified ID for a specific LCO.
      operationId: searchForStudentProgress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentProgressRequest'
      responses:
        200:
          description: Successfully retrieved student progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentLearningContentObjectProgress'
        400:
          description: Bad request
        401:
          description: Authentication failed
        404:
          description: User not found
  /api/v1/student/{userId}/assistance-level:
    get:
      tags:
      - Student Module
      summary: Get assistance level
      description: Get the assistance level for a certain user with the specified ID.
      operationId: getAssistanceLevel
      parameters:
      - $ref: '#/components/parameters/userId'
      responses:
        200:
          description: Successfully retrieved assistance level
          content:
            application/json:
              schema:
                type: object
                description: A measure that defines how much assistance the user with the specified ID wants to receive.
                required:
                - level
                properties:
                  level:
                    $ref: '#/components/schemas/AssistanceLevel'
        400:
          description: Bad request
        401:
          description: Authentication failed
        404:
          description: User not found
    put:
      tags:
      - Student Module
      summary: Set assistance level
      description: Set the assistance level for a certain user.
      operationId: setAssistanceLevel
      parameters:
      - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: A measure that defines how much assistance the user wants to receive.
              required:
              - level
              properties:
                level:
                  $ref: '#/components/schemas/AssistanceLevel'
      responses:
        200:
          description: Successfully set assistance level
        400:
          description: Bad request
        401:
          description: Authentication failed
        404:
          description: User not found
        422:
          description: Unprocessable entity
    x-codegen-request-body-name: body
...
components:
  schemas:
...
    AssistanceLevel:
      type: integer
      description: Schema for representing the level that defines how much assistance is provided to a user that is represented by a value between 0 and 10. The default is the maximal level.
      minimum: 0
      maximum: 10
    LearningPathElement:
      type: object
      description: Schema for representing an element of a learning path.
      required:
      - learningPathElementTarget
      - step
      properties:
        learningPathElementId:
          type: string
          description: The ID of the learning path element that is unique per learning path.
        learningPathElementTarget:
          type: string
          description: The ID of the element that is represented by the learning path element e.g., a content page.
        step:
          type: integer
          minimum: 1
          description: The step number of the element within the corresponding learning path.
        preferred:
          type: boolean
          description: Whether the learning path element is preferred or not.
        learningPathElementProcessedRequirements:
          type: array
          description: A number of learning path elements that have to be processed when this element is processed.
          items:
            $ref: '#/components/schemas/LearningPathElement'
        learningPathElementNotProcessedRequirements:
          type: array
          description: A number of learning path elements that must not have been processed when this element is processed.
          items:
            $ref: '#/components/schemas/LearningPathElement'
    LearningPath:
      type: object
      description: Schema for representing a user's learning path.
      required:
      - title
      - numberOfCurrentAberrations
      properties:
        title:
          type: string
          description: The title of the learning path.
        level:
          type: string
          description: The difficulty level of the learning path.
          # TODO: Add example
        numberOfCurrentAberrations:
          type: integer
          description: The current number of aberrations from this learning path.
          example: 1
        processedLearningPathElements:
          type: array
          description: The already processed learning path elements.
          items:
            $ref: '#/components/schemas/LearningPathElement'
        currentLearningPathElement:
          type: object
          description: The learning path element the user is currently working with.
          allOf:
          - $ref: '#/components/schemas/LearningPathElement'
        successiveLearningPathElements:
          type: array
          description: The learning path elements yet to process.
          items:
            $ref: '#/components/schemas/LearningPathElement'
    StudentModel:
      type: object
      description: Schema for representing a model that encapsulates the knowledge about a certain student.
      required:
      - userId
      properties:
        userId:
          type: string
          description: The unique ID of a user which is also used by the xAPI statements to identify this user.
          example: bb85da8424ffa1eb1c0a328e12f68312c01511ffe0643599c04525902087266c@f269323f-fa99-4102-81fa-2e6ee79d13e8.ilias
        assistanceLevel:
          $ref: '#/components/schemas/AssistanceLevel'
        cooperativeness:
          type: boolean
          description: Whether the user is willing to help others on already solved tasks. The default is false.
        online:
          type: boolean
          description: Whether the user is logged in or not. The default is false.
        experiences:
          type: array
          description: An ordered list of experiences collected for the learner.
          items:
            $ref: '#/components/schemas/Experience'
        learningPathInProgress:
          $ref: '#/components/schemas/LearningPath'
    StudentProgressRequest:
      type: object
      description: Schema for representing a request for the progress of a specific LCO of a specific user.
      required:
      - userId
      properties:
        userId:
          type: string
          description: Unique ID of a user.
          example: bb85da8424ffa1eb1c0a328e12f68312c01511ffe0643599c04525902087266c@f269323f-fa99-4102-81fa-2e6ee79d13e8.ilias
        subLcos:
          type: boolean
          description: Whether the sub LCOs of a LCO should be included in a progress report. The default is false.
          default: false
        lcoId:
          type: string
          description: The unique ID of the LCO for which the progress is requested. This is required if no `objectId` is specified.
          example: 633A587A-B944-4974-850F-DDDE69CECFF1
        objectId:
          type: string
          description: The ID which is used to identify the LCO in an xAPI-statement for which the progress is requested. This is required if no `lcoId` is specified.
          example: http://localhost:8081/goto.php?target=pg__86&client_id=default&h5p_object_id=1&obj_id_lrs=320
    Experience:
      type: object
      description: Schema for representing an experience collected for a learner.
      required:
      - timestamp
      properties:
        timestamp:
          type: string
          description: The timestamp of the experience.
          example: 2023-06-27T10:12:53.000000+02:00
        statementId:
          type: string
          description: The ID of the statement to which the experience relates.
          example: f345c582-cc10-4bab-8554-796ed7261bb9
        objectId:
          type: string
          description: The ID of the object which is referenced in the related statement.
          example: http://localhost:8081/goto.php?target=pg__86&client_id=default&h5p_object_id=1&obj_id_lrs=320
        lcoId:
          type: string
          description: The unique ID of this LCO.
          example: 633A587A-B944-4974-850F-DDDE69CECFF1
        verbId:
          type: string
          description: The ID of the verb which is referenced in the related statement.
          example: http://adlnet.gov/expapi/verbs/answered
        result:
          type: object
          description: The result of the experience, if any, as reported in the statement.
          example: {"score": {"min": 0, "max": 1, "raw": 1, "scaled": 1}, "success": true, "completion": true, "duration": "PT55.41S"}
    StudentLearningContentObjectProgress:
      type: object
      description: Progress of an LCO of a student.
      properties:
        userId:
          type: string
          description: The ID of the user the assistance belongs to.
          example: bb85da8424ffa1eb1c0a328e12f68312c01511ffe0643599c04525902087266c@f269323f-fa99-4102-81fa-2e6ee79d13e8.ilias
        progress:
          type: array
          items:
            $ref: '#/components/schemas/Experience'
        sub_lco_progress:
          type: array
          items:
            type: object
            required:
            - key
            properties:
              key:
                type: string
                description: The identified of the LCO that was experienced.
                example: http://localhost:5173/goto.php?target=lm_90&client_id=default&obj_id_lrs=339
              value:
                type: array
                items:
                  $ref: '#/components/schemas/Experience'
...
    StudentModelList:
      type: object
      description: Schema for representing a container object that encapsulates Student Models.
      required:
      - students
      - totalNumber
      - providedNumber
      - pageNumber
      properties:
        students:
          type: array
          description: The encapsulated Student Models.
          items:
            $ref: '#/components/schemas/StudentModel'
        totalNumber:
          type: integer
          description: The total number of Student Models encapsulated.
          minimum: 0
        providedNumber:
          type: integer
          description: The provided number of Student Models.
          minimum: 0
        pageNumber:
          type: integer
          description: The provided page.
          minimum: 1
...
  parameters:
...
    userId:
      name: userId
      in: path
      description: Unique ID of a user.
      required: true
      schema:
        type: string
        example: bb85da8424ffa1eb1c0a328e12f68312c01511ffe0643599c04525902087266c@f269323f-fa99-4102-81fa-2e6ee79d13e8.ilias
...
    numberOfObjectsPerPage:
      name: objectsPerPage
      in: query
      description: The number of objects that should be delivered per page. If this is not specified, all objects are delivered. This has to be specified when a page number is specified.
      required: false
      schema:
        type: integer
        minimum: 1
    pageNumber:
      name: page
      in: query
      description: The page that should be delivered. The default value is one.
      required: false
      schema:
        type: integer
        minimum: 1
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT